[Optimization][admin] modified some naming and logical sinking
