+++ dinky-admin/src/main/java/org/dinky/common/result/Result.java
@@ -118,4 +118,17 @@ public class Result<T> implements Serializable {
     public static <T> Result<T> errorWithArgs(Status status, Object... args) {
         return new Result<>(status.getCode(), MessageFormat.format(status.getMsg(), args));
     }
+
+    /**
+     * this method returns a successful result with the status
+     *
+     * @param status
+     * @return result code
+     */
+    public static Result<Void> getResultFromStatus(Status status){
+        Result<Void> result = new Result<Void>();
+        result.setCode(status.getCode());
+        result.setMsg(status.getMsg());
+        return result;
+    }
 }
